Cleaned up player creation and substation creation a bit, primarily removing duplicate code
