+++ x/structs/keeper/substation.go
@@ -49,24 +49,32 @@ func (k Keeper) SetSubstationCount(ctx sdk.Context, count uint64) {
 // AppendSubstation appends a substation in the store with a new id and update the count
 func (k Keeper) AppendSubstation(
 	ctx sdk.Context,
-	substation types.Substation,
-) uint64 {
-	// Create the substation
-	count := k.GetSubstationCount(ctx)
+	//substation types.Substation,
+	playerConnectionAllocation uint64,
+    player types.Player,
+) (types.Substation) {
+	substation := types.CreateEmptySubstation()
 
 	// Set the ID of the appended value
-	substation.Id = count
+    substation.SetId(k.GetNextSubstationId(ctx))
 
+    // Setup some Substation details
+    substation.SetOwner(player.Id)
+    substation.SetCreator(player.Creator)
+    substation.SetPlayerConnectionAllocation(playerConnectionAllocation)
+    k.SubstationPermissionAdd(ctx, substation.Id, player.Id, types.SubstationPermissionAll)
+
+
+    // actually commit to the store
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.SubstationKey))
 	appendedValue := k.cdc.MustMarshal(&substation)
 	store.Set(GetSubstationIDBytes(substation.Id), appendedValue)
 
-	// Update substation count
-	k.SetSubstationCount(ctx, count+1)
 
+    // Cache invalidation event
     _ = ctx.EventManager().EmitTypedEvent(&types.EventCacheInvalidation{ObjectId: substation.Id, ObjectType: types.ObjectType_substation})
 
-	return count
+	return substation
 }
 
 // SetSubstation set a specific substation in the store
@@ -501,76 +509,3 @@ func (k Keeper) SubstationDecrementConnectedPlayerCount(ctx sdk.Context, id uint
 
 
 
-
-
-// GetSubstationPermissionIDBytes returns the byte representation of the Substation and player id pair
-func GetSubstationPermissionIDBytes(substationId uint64, playerId uint64) []byte {
-	substationIdString  := strconv.FormatUint(substationId, 10)
-	playerIdString := strconv.FormatUint(playerId, 10)
-
-	return []byte(substationIdString + "-" + playerIdString)
-}
-
-
-func (k Keeper) SubstationGetPlayerPermissionsByBytes(ctx sdk.Context, permissionRecord []byte) (types.SubstationPermission) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.SubstationPermissionKey))
-
-	bz := store.Get(permissionRecord)
-
-	// Substation Capacity Not in Memory: no element
-	if bz == nil {
-		return types.SubstationPermissionless
-	}
-
-	load := types.SubstationPermission(binary.BigEndian.Uint16(bz))
-
-	return load
-}
-
-func (k Keeper) SubstationSetPlayerPermissionsByBytes(ctx sdk.Context, permissionRecord []byte, permissions types.SubstationPermission) (types.SubstationPermission) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.SubstationPermissionKey))
-
-	bz := make([]byte, 8)
-	binary.BigEndian.PutUint16(bz, uint16(permissions))
-
-	store.Set(permissionRecord, bz)
-
-	return permissions
-}
-
-func (k Keeper) SubstationPermissionClearAll(ctx sdk.Context, SubstationId uint64, playerId uint64) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.SubstationPermissionKey))
-	store.Delete(GetSubstationPermissionIDBytes(SubstationId, playerId))
-}
-
-func (k Keeper) SubstationPermissionAdd(ctx sdk.Context, substationId uint64, playerId uint64, flag types.SubstationPermission) types.SubstationPermission {
-    permissionRecord := GetSubstationPermissionIDBytes(substationId, playerId)
-
-    currentPermission := k.SubstationGetPlayerPermissionsByBytes(ctx, permissionRecord)
-    newPermissions := k.SubstationSetPlayerPermissionsByBytes(ctx, permissionRecord, currentPermission | flag)
-	return newPermissions
-}
-
-func (k Keeper) SubstationPermissionRemove(ctx sdk.Context, substationId uint64, playerId uint64, flag types.SubstationPermission) types.SubstationPermission {
-    permissionRecord := GetSubstationPermissionIDBytes(substationId, playerId)
-
-    currentPermission := k.SubstationGetPlayerPermissionsByBytes(ctx, permissionRecord)
-    newPermissions := k.SubstationSetPlayerPermissionsByBytes(ctx, permissionRecord, currentPermission &^ flag)
-	return newPermissions
-}
-
-func (k Keeper) SubstationPermissionHasAll(ctx sdk.Context, substationId uint64, playerId uint64, flag types.SubstationPermission) bool {
-    permissionRecord := GetSubstationPermissionIDBytes(substationId, playerId)
-
-    currentPermission := k.SubstationGetPlayerPermissionsByBytes(ctx, permissionRecord)
-
-	return currentPermission&flag == flag
-}
-
-func (k Keeper) SubstationPermissionHasOneOf(ctx sdk.Context, substationId uint64, playerId uint64, flag types.SubstationPermission) bool {
-    permissionRecord := GetSubstationPermissionIDBytes(substationId, playerId)
-
-    currentPermission := k.SubstationGetPlayerPermissionsByBytes(ctx, permissionRecord)
-
-	return currentPermission&flag != 0
-}
